// src/frontend/src/app/models/cultura-parcela.model.ts
export interface CulturaParcela {
  id?: number;
  anAgricol: number;
  specieCultura: string;
  suprafataCultivataHa: number;
  dataInsamantare?: string;
  dataRecoltare?: string;
  productieHectarKg?: number;
  productieTotalaTone?: number;
  sistemIrigare?: string;
  tipSol?: string;
  culturaPrecedenta?: string;
  parcelaId?: number;
}